Make the Linksys Combo PCMCIA EthernetCard (EC2T) work again; it was broken by the last commit for AX88190 support. There are at least three variants of ne's that have manufacturer id 0x143 (Linksys) + product id 0xc1ab. Each one has its enaddr in a different place.
The matching code for PCMCIA NE2000 is really gross. In particular, the flags
field in ne2000_devs (for NE2000DVF_DL10019 and NE2000_AX88190) is a complete
hack. The driver should really try each method of reading the Ethernet address
(from the ROM, DL10019, and AX88190) and choose the one that generates the
correct checksum.

I plan to fix this later...