This driver supports PC display adapter hardware within the wscons(4) console framework. It doesn't provide direct device driver entry points but makes its functions available via the internal wsdisplay(4) interface.
The pcdisplay driver is indended as a minimal ``catch-all'' driver for
the different kinds of MDA or CGA compatible adapters. It doesn't support
multiple screens, nor colors or font loading.

From NetBSD. Contains routines used by the vga(4) and ega(4) drivers, etc.
This is more MI stuff in prepartion for wscons on i386 and alpha.