This fixes keyboard lock-ups I was getting wiggling the mouse in console mode while using my omnicube KVM switch.
The keyboard lock-ups were caused by the keyboard/mouse interface controller's
input buffer filling up. This was caused because the mouse was
sending data to the computer and the pms driver wasn't draining it.

The driver used to rely on the fact that it could disable the mouse
device when the mouse driver was closed. Disabling the mouse device
prevents it from sending data to the computer.

At the same time pms disabled the mouse device, pms disabled interrupt
notification of the arrival of new data.

This regime doesn't work so well with KVM switches which re-enable the mouse.
In this case, the mouse will send data to the computer but the keyboard/moust
interface controller won't raise an interrupt (because pms disabled
interrupts). pms will never read from the input buffer.