Re-synchronize TSFT after an IBSS merge.
Set up the Tx descriptor ring more safely.  Fix an endianness bug.
"It's a wonder this ever worked."  (Actually, it's not.)

hange the order in which I write the BSSID registers on the ADM8211,
to match the reference driver.  This probably does not make any
functional difference.

Only write the SSID buffer to the ADM8211 SRAM up to the end of
the SSID, not up to the end of the buffer.  Given the (too
conservative?) delays involved, this should save some time when we
join a new network.

From NetBSD (dyoung)