introduce the command control blocks, and the concept that they are managed per port rather than as a big list for the whole controller. the softc has an array of pointers to a struct that manages all the shizz relevant to a particular port. when we map the controllers ports, we'll allocate these as needed.
if the port isnt implemented, we leave its pointer in the softc null and
use that to decide if a drive might be there or not, rather than lookup up
the PI register all the time. a comparison has to null has to be cheaper
than doing bit operations, surely.