Change sis_chip_map() to use pci_find_device() as the original NetBSD code it was based on did/does, now we have it.
This makes detecting sis 96x work properly without having to
do silly things like adding host bridges in the list.  Of
course we wouldn't be in this mess in the first place if sis
designed their chips properly instead of reusing the same
pci id.

Add 964/965 devices while here.

Tested by todd@ on a machine with 965.