The i915 vblank swap ioctl is fundamentally racy.
using it allowed rendering to continue while waiting for a vblank swap,
and often this lead to flickering and rendering a new scene before the
swap. this broke a lot of things.

With the removal of this swap, userland falls back to the old way of
waiting for the vblank then doing the swap itself, this is smooth
enough.

I decided independantly to kill this, but the intel guys recently
concurred.  Comment change comes from Eric Anholt at intel.