Previous code to handle suspend/resume for intagp assumed that everything was unbound before we go down.
Due to software fallbacks from the vt-switched X this is not strictly
true, so save/restore the gtt (which is stored in a BAR and thus
volatile).

i915 and i945 still have some (unrelated) coruption on dzen2 on resume
which I am looking for still (theory: fence registers), but this helps a
LOT of machines.

``commit''  deraadt@. tested by many (and in snaps the last few days.