If running on the original 33C93, reject commands not in groups 0, 1 or 5 if their size is not exactly six bytes, as the chip can't cope with this situation.
Another situation all 33C93 do not cope with very well, is sending stop
commands to targets (such as all sd(4) devices when halting with poweroff) -
it takes a very long time to recover once all targets on the bus have been
powered down, so we need to raise timeouts to unholy values (one test case has
required more than 20 seconds to recover). Not surprising, as this command
is not documented as supported in the chip documentation.