Split the magic to add a connected route for every new IPv4 address added to an interface into two functions: one for route to prefix and one for route to host.
This magic has been added to work around the behavior of rtinit() and
preserve the connected route to prefix when the first address of a
subnet is removed.  Since this "feature" has been present for point-
to-point interfaces too for more than 10 years, this diff does not
change anything in this regard, even if I'm not sure it makes sense.

Tested by benno@